Refactor NotFound
The NotFound component renders the Header component by default, which
works for non-existent endpoints but doesn't work for components that
may need to render the error due to missing data. Split off the error
portion of NotFound into its own component that can be rendered
without the Header.
