+++ frontend/src/components/theme/not-found/index.js
@@ -2,6 +2,11 @@
 
 // This file imports and exports the NotFound component.
 
+import { Error404 } from
+  'components/theme/not-found/not-found-components';
+
 import NotFound from 'components/theme/not-found/not-found';
+
+export { Error404 };
 
 export default NotFound;
