+++ frontend/src/components/theme/not-found/not-found-components/index.js
@@ -1,0 +1,8 @@
+// ./src/components/theme/not-found/not-found-components/index.js
+
+// This file imports and exports the NotFound component.
+
+import Error404 from
+  'components/theme/not-found/not-found-components/error-404';
+
+export { Error404 };
